requirements.txt
flask
flask_sqlalchemy
flask-wtf
pymysql
